Moved configure up
